# src/main/resources/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PDF Generator",
    "description": "Application to create a PDF using an API",
    "contact": {
      "name": "Ewantha Udagama",
      "url": "https://github.com/Ewantha",
      "email": "ewanthau@gmail.com"
    },
    "version": "v0.0.1"
  },
  "externalDocs": {
    "description": "itext Website",
    "url": "https://itextpdf.com/"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/document/pdf": {
      "post": {
        "tags": [
          "document-controller"
        ],
        "summary": "Create a PDF Document",
        "operationId": "generateDocument",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateDocumentRequest": {
        "required": [
          "DocumentName"
        ],
        "type": "object",
        "properties": {
          "DocumentName": {
            "type": "string",
            "description": "Name of the PDF document",
            "example": "HelloWorld"
          },
          "Content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentContent"
            }
          }
        }
      },
      "DocumentContent": {
        "required": [
          "Type"
        ],
        "type": "object",
        "properties": {
          "Type": {
            "type": "string",
            "description": "Type of document element to be created.",
            "enum": [
              "Paragraph",
              "Anchor",
              "Image",
              "Table",
              "Whitespace"
            ]
          },
          "Font": {
            "$ref": "#/components/schemas/ElementFont"
          },
          "Text": {
            "type": "string",
            "description": "Input text.",
            "example": "Lorem ipsum..."
          },
          "Reference": {
            "type": "string",
            "description": "Only for anchor elements and image elements. Set url or image source.",
            "example": "www.google.com"
          },
          "Scale": {
            "type": "number",
            "description": "Only for image and table elements.",
            "format": "float",
            "example": 1.2
          },
          "Columns": {
            "type": "integer",
            "description": "Only for table elements.",
            "format": "int32",
            "example": 2
          },
          "Data": {
            "type": "array",
            "description": "Only for table elements.",
            "items": {
              "type": "string",
              "description": "Only for table elements."
            }
          },
          "Header": {
            "type": "array",
            "description": "Only for table elements.",
            "items": {
              "type": "string",
              "description": "Only for table elements."
            }
          },
          "HeaderFont": {
            "$ref": "#/components/schemas/ElementFont"
          }
        }
      },
      "DocumentFontColor": {
        "type": "object",
        "properties": {
          "Red": {
            "type": "integer",
            "format": "int32"
          },
          "Green": {
            "type": "integer",
            "format": "int32"
          },
          "Blue": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ElementFont": {
        "type": "object",
        "properties": {
          "Type": {
            "type": "string"
          },
          "Size": {
            "type": "integer",
            "format": "int32"
          },
          "Style": {
            "type": "integer",
            "format": "int32"
          },
          "Color": {
            "$ref": "#/components/schemas/DocumentFontColor"
          }
        },
        "description": "Only for table elements."
      }
    }
  }
}